Intermediate shutdown task 6
Использовал простой путь для работы с апи.Для полного завернеия задания необзодимо слияние с 5тым таском.
