+++ ConsoleApp2/TreeNode.cs
@@ -1,0 +1,70 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace ConsoleApp2
+{
+    public class TreeNode
+    {
+        public TreeNode(char data)
+        {
+            Data = data;
+        }
+
+        //данные
+        public char Data { get; set; }
+
+        //левая ветка дерева
+        public TreeNode Left { get; set; }
+
+        //правая ветка дерева
+        public TreeNode Right { get; set; }
+
+        //добавление узла в дерево
+        public void Insert(TreeNode node)
+        {
+            if (node.Data < Data)
+            {
+                if (Left == null)
+                    Left = node;
+                else
+                    Left.Insert(node);
+            }
+            else
+            {
+                if (Right == null)
+                    Right = node;
+                else
+                    Right.Insert(node);
+            }
+        }
+
+        public char[] Transform(List<char> elements = null)
+        {
+            if (elements == null)
+                elements = new List<char>();
+
+            if (Left != null)
+                Left.Transform(elements);
+
+            elements.Add(Data);
+
+            if (Right != null)
+                Right.Transform(elements);
+
+            return elements.ToArray();
+        }
+        public static char[] TreeSort(char[] array)
+        {
+            var treeNode = new TreeNode(array[0]);
+            for (int i = 1; i < array.Length; i++)
+            {
+                treeNode.Insert(new TreeNode(array[i]));
+            }
+
+            return treeNode.Transform();
+        }
+    }
+}
